1) View/ Dlg 에 GetDocument() 포인터 읽어서 데이터 처리
